+++ kernel/src/model/sample/mod.rs
@@ -7,5 +7,5 @@ pub struct Sample {
     pub id: SampleId,
     pub name: String,
     pub email: String,
-    pub age: i32,
+    pub age: i16,
 }
